refactor: FileCleanupScheduler를 adapter 레이어로 이동
