+++ src/main/java/com/dreamteam/alter/application/file/usecase/CleanupOrphanFiles.java
@@ -1,9 +1,9 @@
 package com.dreamteam.alter.application.file.usecase;
 
+import com.dreamteam.alter.application.file.FileDeleteService;
 import com.dreamteam.alter.domain.file.entity.File;
 import com.dreamteam.alter.domain.file.port.inbound.CleanupOrphanFilesUseCase;
 import com.dreamteam.alter.domain.file.port.outbound.FileQueryRepository;
-import com.dreamteam.alter.domain.file.port.outbound.S3Client;
 import jakarta.annotation.Resource;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,8 +20,8 @@ public class CleanupOrphanFiles implements CleanupOrphanFilesUseCase {
 
     private final FileQueryRepository fileQueryRepository;
 
-    @Resource(name = "s3Client")
-    private final S3Client s3Client;
+    @Resource(name = "fileDeleteService")
+    private final FileDeleteService fileDeleteService;
 
     @Override
     @Transactional
@@ -29,15 +29,18 @@ public class CleanupOrphanFiles implements CleanupOrphanFilesUseCase {
         LocalDateTime threshold = LocalDateTime.now().minusHours(24);
         List<File> orphanFiles = fileQueryRepository.findOrphanFiles(threshold);
 
+        int successCount = 0;
+        int failureCount = 0;
         for (File file : orphanFiles) {
             try {
-                s3Client.delete(file.getStoredKey(), file.getBucketType());
-                file.markDeleted();
+                fileDeleteService.delete(file);
+                successCount++;
             } catch (Exception e) {
                 log.error("Failed to cleanup orphan file id={}, key={}", file.getId(), file.getStoredKey(), e);
+                failureCount++;
             }
         }
 
-        log.info("Orphan file cleanup completed. processed={}", orphanFiles.size());
+        log.info("Orphan file cleanup completed. total={}, success={}, failed={}", orphanFiles.size(), successCount, failureCount);
     }
 }
